services/memory_builder/requirements.txt
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0